pyproject.toml
[project]
name = "cafetwin"
version = "0.1.0"
description = "CafeTwin / SimCafe hackathon MVP"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "logfire[fastapi]>=3.0.0",
    "pydantic>=2.7.0",
    "pydantic-ai>=1.16.0",
    "python-dotenv>=1.0.1",
    "uvicorn[standard]>=0.30.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.6.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
